filter functionality kinda added
A non-functional filters.js has been added and a button is at the bottom of the page can be used to test it. Results can be seen through the server console.log messages. The sql command should be corrected but it hasn't actually been tested because the database is not up at the moment.

The npm module async is now required for this code to function. It can be installed using "npm install async"